refactor(infra): CDK stack code cleanup
